Refactored the sqs handlers to follow the pattern that has been set for the api, also updated the eslintrc
