+++ .eslintrc
@@ -29,12 +29,13 @@
     "unicorn/no-abusive-eslint-disable": "error",
     "max-params": ["error", 5],
     "max-depth": ["error", 3],
+    "max-len": ["error", 150],
     "promise/always-return": "error",
     "promise/catch-or-return": "error",
     "promise/param-names": "error",
     "promise/no-return-wrap": "error",
     "no-unused-vars": ["error", { "argsIgnorePattern": "next", "ignoreRestSiblings": true }],
-    "no-param-reassign": ["error", { "props": true, "ignorePropertyModificationsFor": ["result", "res"] }],
+    "no-param-reassign": ["error", { "props": true, "ignorePropertyModificationsFor": ["result", "res", "context"] }],
     "newline-per-chained-call": ["error", { "ignoreChainWithDepth": 2 }]
   }
 }
